short

+++ PacManDataOriented/src/main/java/de/amr/games/pacman/core/Ghost.java
@@ -20,7 +20,7 @@ public class Ghost extends Creature {
 	public boolean locked;
 	public boolean enteringHouse;
 	public boolean leavingHouse;
-	public int bounty;
+	public short bounty;
 	public int dotCounter;
 	public byte elroyMode;
